1. Added dependency injection 2. Added lazy initialization 3. Added regex to email
