+++ WEB/Controllers/AccountController.cs
@@ -5,15 +5,23 @@ using System.Linq;
 using System.Web;
 using System.Web.Http.Results;
 using System.Web.Mvc;
+using WEB.Interfaces;
 using WEB.Models;
 using WEB.ShopFromAListLogic;
+using WEB.Top5Logic;
 
 namespace WEB.Controllers
 {
     public class AccountController : Controller
     {
-        // GET: Account
         List<string> list = new List<string>();
+        private readonly IUserAccountDbContext _context;
+
+        public AccountController(IUserAccountDbContext context)
+        {
+            _context = context;
+        }
+
         public ActionResult Index()
         {
             return View();
@@ -29,18 +37,13 @@ namespace WEB.Controllers
         {
             if (ModelState.IsValid)
             {
-                using (UserAccountDbContext db = new UserAccountDbContext())
-                {
-                    db.userAccount.Add(user);
-                    db.SaveChanges();
-                }
+                _context.userAccount.Add(user);
+                _context.SaveChanges();
                 ModelState.Clear();
                 ViewBag.Message = user.FirstName + " " + user.LastName + " with username: " + user.Username + " succesfully registered!";
-
-
             }
 
-            return View();
+            return RedirectToAction("Login");
         }
 
         public ActionResult Login()
@@ -51,24 +54,20 @@ namespace WEB.Controllers
         [HttpPost]
         public ActionResult Login(UserAccount user)
         {
-            using (UserAccountDbContext db = new UserAccountDbContext())
+            var usr = _context.userAccount.Where(u => u.Username == user.Username && u.Password == user.Password).FirstOrDefault();
+
+            if (usr != null)
             {
-                var usr = db.userAccount.Where(u => u.Username == user.Username && u.Password == user.Password).FirstOrDefault();
-                
-                if (usr != null)
-                {
-                    
-                    Session["UserID"] = usr.UserID.ToString();
-                    Session["Username"] = usr.Username.ToString();
-                    /*db.userAccount.RemoveRange(db.userAccount);
-                    db.SaveChanges();*/
-                    return RedirectToAction("Dashboard");
-                }
-                else
-                {
-                    ModelState.AddModelError("", "Bad Login Credentials");
-                    return View();
-                }
+                Session["UserID"] = usr.UserID.ToString();
+                Session["Username"] = usr.Username.ToString();
+                _context.userAccount.RemoveRange(_context.userAccount);
+                _context.SaveChanges();
+                return RedirectToAction("Dashboard");
+            }
+            else
+            {
+                ModelState.AddModelError("", "Bad Login Credentials");
+                return View();
             }
         }
 
@@ -87,13 +86,10 @@ namespace WEB.Controllers
         public ActionResult StoreList()
         {
             PurchaseList purchaseList = new PurchaseList();
-            using (UserAccountDbContext db = new UserAccountDbContext())
-            {
-                purchaseList.listOfProducts = db.purchasedItem.ToList<PurchasedItem>();
-            }
+            purchaseList.listOfProducts = _context.purchasedItem.ToList<PurchasedItem>();
             return View(purchaseList);
         }
-        
+
         [HttpPost]
         public ActionResult StoreList(PurchaseList purchaseList)
         {
@@ -105,7 +101,7 @@ namespace WEB.Controllers
                 {
                     list.Add(item.ItemName);
                 }
-                FromList fromList = new FromList(list, DateTime.Now.AddMonths(-1));
+                FromList fromList = new FromList(_context, list, DateTime.Now.AddMonths(-1));
                 ViewBag.rezult = fromList.ReturnStoreName();
                 return View("StoreList");
             }
@@ -114,33 +110,41 @@ namespace WEB.Controllers
                 ViewBag.rezult = "Please some check boxes to add items to your list";
                 return View("StoreList");
             }
-           
+
         }
         public ActionResult Top5()
         {
-            using (UserAccountDbContext db = new UserAccountDbContext())
+
+            /*_context.purchasedItem.Add(new PurchasedItem
             {
-                /*db.purchasedItem.Add(new PurchasedItem
-                {
-                    Date = new DateTime(2017, 11, 18),
-                    ItemName = "Pienas",
-                    Price = 0.99,
-                    ShopName = "Maxima",
-                    UserId = Convert.ToInt32(Session["UserID"]),
-                    IsChecked = false
-                });
-                db.SaveChanges();*/
-            }
-            var items = new Top5(Convert.ToInt32(Session["UserID"]));
+                Date = new DateTime(2017, 11, 18),
+                ItemName = "Juoda Duona",
+                Price = 0.89,
+                ShopName = "Iki",
+                UserId = Convert.ToInt32(Session["UserID"]),
+                IsChecked = false
+            });
+            _context.SaveChanges();*/
+
+            var items = new Top5(Convert.ToInt32(Session["UserID"]), _context);
             ViewBag.Warning = items.Warning;
-            ViewBag.Shop = items.Recommendation.Key;
-            ViewBag.Price = items.Recommendation.Value; 
+
+            if (items.Recommendation.Key != null)
+            {
+                ViewBag.Shop = items.Recommendation.Key;
+            }
+            else
+            {
+                ViewBag.Shop = "";
+            }
+
+            ViewBag.Price = items.Recommendation.Value;
             return View(items.Items);
         }
         public ActionResult Logout()
         {
             Session.Abandon();
             return RedirectToAction("Login");
-        }      
+        }
     }
 }
