+++ WEB/Controllers/ImageScanController.cs
@@ -40,7 +40,6 @@ namespace WEB.Controllers
                     }
                     image = memoryStream.ToArray();
                 }
- 
                 return View("Index", (object)GoogleOCR.GetGoogleOCR().DoRecognition(image));
             }
             else
