+++ WEB/Global.asax.cs
@@ -7,6 +7,12 @@ using System.Web.Routing;
 using System.Web.Security;
 using System.Web.SessionState;
 using System.Web.Http;
+using SimpleInjector;
+using SimpleInjector.Integration.Web;
+using SimpleInjector.Integration.Web.Mvc;
+using System.Reflection;
+using WEB.Models;
+using WEB.Interfaces;
 
 namespace WEB
 {
@@ -14,11 +20,28 @@ namespace WEB
     {
         void Application_Start(object sender, EventArgs e)
         {
-            // Code that runs on application startup
             AreaRegistration.RegisterAllAreas();
             GlobalConfiguration.Configure(WebApiConfig.Register);
             RouteConfig.RegisterRoutes(RouteTable.Routes);
+
+            // Create the container as usual.
+            var container = new Container();
+            container.Options.DefaultScopedLifestyle = new WebRequestLifestyle();
+
+            // Register your types, for instance:
+            container.Register<IUserAccountDbContext, UserAccountDbContext>(Lifestyle.Scoped);
+
+            // This is an extension method from the integration package.
+            container.RegisterMvcControllers(Assembly.GetExecutingAssembly());
+
+            // This is an extension method from the integration package as well.
+            container.RegisterMvcIntegratedFilterProvider();
+
+            container.Verify();
+
+            DependencyResolver.SetResolver(new SimpleInjectorDependencyResolver(container));
+            // Code that runs on application startup
         }
-        
+
     }
 }
