+++ WEB/Models/UserAccount.cs
@@ -10,21 +10,29 @@ namespace WEB.Models
     public class UserAccount
     {
         [Key]
+
         public int UserID { get; set; }
         [Required(ErrorMessage = "Username is required!")]
         public string Username { get; set; }
+
         [Required(ErrorMessage = "Name is required!")]
         public string FirstName { get; set; }
+
         [Required(ErrorMessage = "Last Name is required!")]
         public string LastName { get; set; }
+
         [Required(ErrorMessage = "Email is required!")]
+        [RegularExpression(@"^([\w\.\-]+)@([\w\-]+)((\.(\w){2,3})+)$", ErrorMessage = "Invalid Email address!")]
         public string Email { get; set; }
+
         [Required(ErrorMessage = "Password is required!")]
         [DataType(DataType.Password)]
         public string Password { get; set; }
+
         [Required(ErrorMessage = "Password Confirmation is required!")]
         [DataType(DataType.Password)]
         public string ConfirmPassword { get; set; }
+
         public virtual List<PurchasedItem> PurchasedItems { get; set; }
     }
 }
