+++ WEB/Models/UserAccountDbContext.cs
@@ -3,14 +3,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 using System.Data.Entity;
+using WEB.Interfaces;
 
 namespace WEB.Models
 {
-    public class UserAccountDbContext : DbContext
+    public class UserAccountDbContext : DbContext, IUserAccountDbContext
     {
         public DbSet<UserAccount> userAccount { get; set; }
         public DbSet<PurchasedItem> purchasedItem { get; set; }
-        public DbSet<Top5> top5Items { get; set; }
         public DbSet<Top5Item> top5Item { get; set; }
 
         protected override void OnModelCreating(DbModelBuilder modelBuilder)
