+++ WEB/Views/Map/Index.cshtml
@@ -34,7 +34,7 @@
 
         var request = {
             location: currentLoc,
-            radius: 20000,
+            radius: 800,
             name: [shop]
         };
         infowindow = new google.maps.InfoWindow();
