+++ WEB/packages.config
@@ -34,6 +34,9 @@
   <package id="MoreLinq.Source.MoreEnumerable.MinBy" version="1.0.1" targetFramework="net461" />
   <package id="NETStandard.Library" version="1.6.0" targetFramework="net461" />
   <package id="Newtonsoft.Json" version="10.0.2" targetFramework="net461" />
+  <package id="SimpleInjector" version="4.0.12" targetFramework="net461" />
+  <package id="SimpleInjector.Integration.Web" version="4.0.12" targetFramework="net461" />
+  <package id="SimpleInjector.Integration.Web.Mvc" version="4.0.12" targetFramework="net461" />
   <package id="System.AppContext" version="4.1.0" targetFramework="net461" />
   <package id="System.Collections" version="4.0.11" targetFramework="net461" />
   <package id="System.Collections.Concurrent" version="4.0.12" targetFramework="net461" />
